requirements.txt
asgiref==3.7.2
Django==5.0
django-ckeditor==6.7.0
django-extensions==3.2.3
django-js-asset==2.1.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pydot==2.0.0
pyparsing==3.1.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
